+++ src/Sulu/Component/Content/Document/Subscriber/ResourceSegmentSubscriber.php
@@ -21,6 +21,7 @@ use Sulu\Component\Content\Metadata\PropertyMetadata;
 use Sulu\Component\Content\Types\Rlp\Strategy\RlpStrategyInterface;
 use Sulu\Component\DocumentManager\DocumentManagerInterface;
 use Sulu\Component\DocumentManager\Event\AbstractMappingEvent;
+use Sulu\Component\DocumentManager\Event\CopyEvent;
 use Sulu\Component\DocumentManager\Event\MoveEvent;
 use Sulu\Component\DocumentManager\Event\PersistEvent;
 use Sulu\Component\DocumentManager\Events;
@@ -80,6 +81,7 @@ class ResourceSegmentSubscriber implements EventSubscriberInterface
             // hydrate should happen afterwards
             Events::HYDRATE => ['handleHydrate', -200],
             Events::MOVE => ['moveRoutes', -128],
+            Events::COPY => ['copyRoutes', -128],
         ];
     }
 
@@ -176,34 +178,23 @@ class ResourceSegmentSubscriber implements EventSubscriberInterface
      */
     public function moveRoutes(MoveEvent $event)
     {
-        $document = $event->getDocument();
+        $this->recreateRoutes($event->getDocument());
+    }
 
-        if (!$document instanceof ResourceSegmentBehavior) {
-            return;
-        }
-
-        $locales = $this->documentInspector->getLocales($document);
-        $webspaceKey = $this->documentInspector->getWebspace($document);
-        $childUuid = $this->documentInspector->getUuid($document);
-        $parentUuid = $this->documentInspector->getUuid($this->documentInspector->getParent($document));
-
-        foreach ($locales as $locale) {
-            $localizedDocument = $this->documentManager->find($childUuid, $locale);
-
-            if ($localizedDocument->getRedirectType() !== RedirectType::NONE) {
-                continue;
-            }
-
-            $parentPart = $this->rlpStrategy->loadByContentUuid($parentUuid, $webspaceKey, $locale);
-            $childPart = $this->rlpStrategy->loadByContentUuid($childUuid, $webspaceKey, $locale);
-            $childPart = $this->rlpStrategy->getChildPart($childPart);
-
-            $localizedDocument->setResourceSegment(
-                $this->rlpStrategy->generate($childPart, $parentPart, $webspaceKey, $locale)
-            );
-
-            $this->documentManager->persist($localizedDocument, $locale);
-        }
+    /**
+     * Copy the routes for all localization of the document in the event.
+     *
+     * @param CopyEvent $event
+     */
+    public function copyRoutes(CopyEvent $event)
+    {
+        $this->recreateRoutes(
+            $event->getDocument(),
+            $this->documentManager->find(
+                $event->getCopiedPath(),
+                $this->documentInspector->getLocale($event->getDocument())
+            )
+        );
     }
 
     /**
@@ -253,5 +244,51 @@ class ResourceSegmentSubscriber implements EventSubscriberInterface
     private function persistRoute(ResourceSegmentBehavior $document)
     {
         $this->rlpStrategy->save($document, null);
+    }
+
+    /**
+     * Recreates the routes for the destination document with the data from the source document.
+     *
+     * Note that both documents can be the same (e.g. happening on a move). If the second parameter is omitted it has
+     * the same value as the first one.
+     *
+     * @param object $sourceDocument
+     * @param object $destinationDocument
+     */
+    private function recreateRoutes($sourceDocument, $destinationDocument = null)
+    {
+        $destinationDocument = $destinationDocument ?: $sourceDocument;
+
+        if (!$sourceDocument instanceof ResourceSegmentBehavior
+            || !$destinationDocument instanceof ResourceSegmentBehavior
+        ) {
+            return;
+        }
+
+        $locales = $this->documentInspector->getLocales($destinationDocument);
+        $webspaceKey = $this->documentInspector->getWebspace($destinationDocument);
+        $sourceUuid = $this->documentInspector->getUuid($sourceDocument);
+        $destinationUuid = $this->documentInspector->getUuid($destinationDocument);
+        $destinationParentUuid = $this->documentInspector->getUuid(
+            $this->documentInspector->getParent($destinationDocument)
+        );
+
+        foreach ($locales as $locale) {
+            $localizedDocument = $this->documentManager->find($destinationUuid, $locale);
+
+            if ($localizedDocument->getRedirectType() !== RedirectType::NONE) {
+                continue;
+            }
+
+            $parentPart = $this->rlpStrategy->loadByContentUuid($destinationParentUuid, $webspaceKey, $locale);
+            $childPart = $this->rlpStrategy->loadByContentUuid($sourceUuid, $webspaceKey, $locale);
+            $childPart = $this->rlpStrategy->getChildPart($childPart);
+
+            $localizedDocument->setResourceSegment(
+                $this->rlpStrategy->generate($childPart, $parentPart, $webspaceKey, $locale)
+            );
+
+            $this->documentManager->persist($localizedDocument, $locale);
+        }
     }
 }